+++ src/ol/render/canvas/ImageBuilder.js
