+++ src/ol/render/canvas/LineStringBuilder.js
