+++ src/ol/render/canvas/PolygonBuilder.js
