+++ src/ol/render/canvas/TextBuilder.js
