+++ test/spec/ol/renderer/canvas/replay.test.js
@@ -7,10 +7,10 @@ import MultiPoint from '../../../../../src/ol/geom/MultiPoint.js';
 import MultiPolygon from '../../../../../src/ol/geom/MultiPolygon.js';
 import Point from '../../../../../src/ol/geom/Point.js';
 import Polygon from '../../../../../src/ol/geom/Polygon.js';
-import CanvasLineStringReplay from '../../../../../src/ol/render/canvas/LineStringReplay.js';
-import CanvasPolygonReplay from '../../../../../src/ol/render/canvas/PolygonReplay.js';
-import CanvasReplay from '../../../../../src/ol/render/canvas/Replay.js';
-import CanvasReplayGroup from '../../../../../src/ol/render/canvas/ReplayGroup.js';
+import CanvasLineStringReplay from '../../../../../src/ol/render/canvas/LineStringBuilder.js';
+import CanvasPolygonReplay from '../../../../../src/ol/render/canvas/PolygonBuilder.js';
+import CanvasReplay from '../../../../../src/ol/render/canvas/InstructionsBuilder.js';
+import CanvasReplayGroup from '../../../../../src/ol/render/canvas/InstructionsGroupBuilder.js';
 import {renderFeature} from '../../../../../src/ol/renderer/vector.js';
 import Fill from '../../../../../src/ol/style/Fill.js';
 import Stroke from '../../../../../src/ol/style/Stroke.js';
